requirements.txt
colorama
discord.py